authn: Gracefully handle failures deserializing the user from a session
Unexpected errors during deserializeUser() are reported to the logs but
otherwise ignored.  The session remains intact but no user is logged in
for the request.  This behaviour is nicer to users as bugs in session
handling mean they can still use the public portions of the site rather
than get a wall of errors.

In contrast to destroying the session on errors or removing the
serialized user (by returning a null user instead of an error), leaving
the session intact means that it's recoverable after deploying a bug fix
or rolling back to known good versions of the code.  It also means that
incompatible sessions between different versions of the codebase (e.g.
next.nextstrain.org and nextstrain.org) can co-exist without causing
issues for one another.