Include raster versions (PNG) of the logos
For convenience in places where SVGs aren't supported.  I found myself
doing this enough recently that I didn't care to do it again in the
future.  That said, these are high resolution (300 ppi), which may not
be suitable for all places, but at least I don't have to remember the
cairosvg commands.