tests: Move Jest config out of package.json and into a separate file
It's more obvious where the Jest config is when in a separate file and
the JS config file supports comments which the JSON file does not.

Motivated by having just accidentally overridden the Jest config in
package.json without realizing it by introducing a new jest.config.js
with an upcoming change, which broke tests because BASE_URL wasn't
defined.  When I went to move the change into package.json, I realized I
couldn't preserve the explanatory comment, which is helpful to
understand why the change is necessary.