docs: Document sessions
I've been waist-deep in our session handling recently (in this branch
and others before it) and it seemed like a good idea to pull all this
disparate info into one canonical place.  Much of it was previously
written down _somewhere_, but it was scattered across code comments,
commit messages, GitHub reviews, Slack, and more… not to mention the
stuff only in my head.  I'm sure I'll thank myself later!  This isn't
all there is to say about sessions, but it's enough for now I think.