Wait for clients to finish sending the request body before responding with an error
Heroku's routing/proxy layer requires this behaviour, returning a 503
for a server-closed socket if we error early, as do some other clients
like the Python "requests" package.  curl, however, does not because it
stops sending once it receives a response instead of trying to send to a
closed socket.

This has the downside of dragging out requests that fail early (e.g. due
to authorization) while we wait for the body that will never be used to
come over the wire, but there's not much we can do to avoid that given
client behaviour.

Resolves <https://github.com/nextstrain/nextstrain.org/issues/564>.