Support Expect: 100-continue requests with 100 Continue responses
This allows clients that support this expectation, like curl, to avoid
sending the request body when it would only be discarded by the server
anyhow and get a quicker error response.  curl is likely to be a
minority of clients for now, but hopefully in time we can switch the
HTTP client used by the Nextstrain CLI to one that supports this
expectation.

On Heroku, we need to enable the per-app http-end-to-end-continue
feature¹ to permit this expectation through Heroku's routing/proxy
layer.  If the feature is not enabled, the behaviour will be as before
this commit.

Related to <https://github.com/nextstrain/nextstrain.org/issues/564>.

¹ https://devcenter.heroku.com/articles/http-routing#enabling-100-continue-support