charon/getDataset: Only fallback to v1 (meta + tree) JSONs when the request is for a v2 (main) JSON
It makes no sense to look for meta + tree when the request is for a
sidecar, e.g. root-sequence.  This bug was recently introduced in
"charon: Use new "sources" handlers for getDataset, getNarrative, and
getSourceInfo requests" (62d17e1).

The outcome of the bug is that we make more requests than necessary and
could potentially find a meta + tree to send back for a root-sequence
request, though this is unlikely to happen and requires a fairly narrow
and unusual set of circumstances.