charon/getDataset: Only fallback to v1 (meta + tree) if v2 (main) wasn't found
Don't fallback to v1 if the error fetching v2 was something else, like
an authorization or connection error.

Crucially, this ensures that the authz check of Read on the Dataset
(performed via sendDatasetSubresource) isn't bypassed/ignored for v1
datasets.  This couldn't currently result in any actual unauthorized
access, however, because the authz check on the Source was still
performed earlier and in all existing policies anyone with Read on a
Source also has Read on the contained Datasets.