Vendor @awaitjs/express in advance of fixing its buggy next() handling
This is such a small library that it seems simpler to vendor it than go
through an upstream patch cycle.  This is especially true since though
its next() handling is buggy in my eyes since it's a departure from
Express, the behaviour seems to be as-desired by the author (based on
PRs, issue comments, etc).

Copied unmodified for now, so our modifications are easily visible in
version control history.