Fix two bugs in wrappers for async Express handlers (formerly @awaitjs/express)
Only call next() on error; never on success.  It's the job of the
handler itself to call next() (if it's middleware) or send a response
(if it's a terminal handler).  This matches Express' model for request
handling instead of subverting it and fixes an apparent race we observed
on res.headersSent to call next() on success.

The race was causing mixed up request handling that manifested as either
503s (from Heroku's routing layer) or 404s (from our catchall routes)
depending on socket conditions for what should have been 304 responses.
These were most noticeable for /charon/getDataset after the deploy of
"charon: Use new "sources" handlers for getDataset, getNarrative, and
getSourceInfo requests" (62d17e1), but the issue existed before that and
was also reproducible for the equivalent RESTful API endpoints.  It was
only because we didn't have logged-in browser clients with a cache
hitting those RESTful API endpoints that we didn't notice.

Errors are also no longer discarded silently when res.headersSent is
already true but instead still passed outwards back into Express.  They
may still ultimately be _unactionable_ if the response has already
started, but they can at least be logged by the app's error handler
instead of being swallowed silently.  This results in us now noticing a
previously-swallowed premature closure error, which I'll address in a
subsequent commit.

My details on the circumstances and troubleshooting that led to this are
in Slack¹ and my notes.²

¹ https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1657144797720099
² https://github.com/tsibley/blab-standup/blob/master/2022-07-14.md