Ignore premature closure errors when piping an upstream response into our response
When we send a fixed-length response (i.e. with Content-Length and not a
chunked transfer encoding), clients will typically close the connection
immediately after receiving the expected number of bytes.  This causes
`res` to emit a "close` event, which pipeline() notices happens before a
"finish" event, so it emits a premature closure error for the writable
side.¹  This is more of a warning that something *may* be wrong, but in
this case nothing really is wrong and even if it was, there's nothing we
can do about it anyway.  Clients can always close the connection when
they want, even before we write all the data we intend to.

This does also ignore premature closure of the *readable* side of the
pipeline, `upstreamRes.body`, but we can't distinguish between the two
when using pipeline().  In any case, having already sent some data to
the client, there's again nothing to do to recover from that case
anyhow.

We could also avoid the error by setting up pipe()s ourselves, but only
because we'd be omitting the comprehensive error handling that
pipeline() implements which catches this subtle edge case.  I'd rather
have the better, broader error handling with a narrow ignore than narrow
error handling.

¹ https://nodejs.org/docs/latest-v14.x/api/errors.html#errors_err_stream_premature_close