SC2 forecasts page: Add sections for result panels
Split into four sections based on internal discussion on Slack¹ of which
static images should be hosted for MVP: Variant Rt, Estimated Variant
Frequencies, Estimated Variant Cases, and Relative Growth Advantage.

Each section has its own toggles for data provenance and geo-resolution
so that the image displayed for each section can be controlled
independently. I did it this way so that the user does not have to
scroll up to the top of the page to change display options.

Currently links to images hosted in the blab/rt-from-frequency-dynamics
repo or generic placeholder images, but future commits will update this
page to pull the images from files hosted on S3.

¹ https://bedfordlab.slack.com/archives/C036L7V615F/p1660088125125799?thread_ts=1660085748.766369&cid=C036L7V615F