Move code handling upstream URL requests to its own file
This allows for reuse later with other endpoints.

DELETE: Extract the logic of creating multiple DELETE requests in
deleteResource to its own function.

GET: Extract the call to proxyResponseBodyFromUpstream in
sendSubresource to its own function.

PUT: Extract the logic of creating a PUT request in receiveSubresource
to its own function. This requires a new type, upstreamUrlExtractor, to
be used since the URL for these requests are dependent on the method and
headers, which are defined in the new function.