utils/scripts: Remove erroneous assignments to process.stdout and .stderr
Avoids the following kind of error when setupConsole() is invoked by the
CJS → ESM conversion of scripts/migrate-groups:

    TypeError: Cannot set property stdout of #<process> which has only a getter

Even prior to the CJS → ESM conversion, these assignments never worked
(!) because the process.stdout and .stderr properties are read-only
(i.e. define only a getter without a setter).  We didn't notice,
however, because such an assignment is not an error unless you're in
strict-mode, which wasn't enabled here.  The conversion from CJS → ESM
revealed the error because ESM usage implies strict-mode.

IIRC, these lines were a "dot the is and cross the ts" sort of change
hoping to catch anything that directly called process.stdout.write() or
process.stderr.write().  The main functionality of setupConsole() is and
was covered by the re-assignment of the global "console" variable.