aws: Introduce a Terraform configuration for managing resources
Terraform is essentially an industry standard at this point and is the
most boring choice for infra management.  I expect our usage to stay
fairly light for a long time but still provide a better measure of
robustness and version control to our AWS resource management than we
have now (which is not much).  The plan is to start with a few core
resources and then over time incrementally bring more of our
manually-managed resources into the fold.

While the desire for Terraform or something of its ilk has been
simmering for a long time, the catalyzing factor here motivating this
work is the desire/need to set up a separate dev/test Cognito user pool
separate from our production one.