Use stand-alone forecasts viz app
The viz app development has happened in https://github.com/nextstrain/forecasts-ncov/pull/8.

I tried many steps to avoid having another app in our build steps, but
ultimately this was the cleanest! Specifically

* our version of gatsby is not compatible with d3 (specifcially the d3
libraries which use DOM selections). There are a few random GitHub
issue comments indicating a potential solution, but after spending an
afternoon on it I decided this wasn't worth it. Note that (a) we're on
Gatsby v2 still and (b) we want to move off Gatsby.

* I tried using rollup to produce a bundle which Gatsby could import.
Gatsby (perhaps as expected) tries to transpile this which we don't
want. We could imagine having a specific version of the app which
renders (in an IIFE) into a specifc element, and importing this in a
<script> tag, but I didn't explore this.

* I tried cloning / copying the viz source into nextstrain.org and
importing certain static-site components (e.g. header, footer) but this
is not possible in create-react app.

* I tried the above, but copying the (static-site) files (header, footer)
however the dependencies go deep and this felt like a bad road to go down!

Finally I settled on this (temporary) solution. See `./build.sh` for
how to build. The viz app is intentionally designed to use custom asset
paths so that the [nextstrain.org] routing is straightforward - thanks
@tsibley for review comments here!