team: Regularize avatar container widths
I really didn't like the jagged avatar placement in the previous version. Especially with blurb text I much prefer a gridded layout. This gives a simple grid of 3 wide in large screens that goes to 1 wide on mobile. Previous version was not mobile friendly.

If we don't use blurb's for alumni (which we may want to do), then I can see having a width of 4 for the alumni section, but I'll leave this for another commit.