endpoints/cli: Support downloading standalone installation archives from PR builds
This new endpoint allows the standalone installer to install not just
released versions but also the CI builds produced for arbitrary PRs.
That's very helpful for development and testing of PRs.  With this new
endpoint, for example, we can run:

    curl -fsSL --proto '=https' https://nextstrain.org/cli/installer/linux \
        | DESTINATION=/tmp/cli bash -s pr-build/243

to install /tmp/cli/nextstrain from the last successful CI build for:

    https://github.com/nextstrain/cli/pulls/243

This uses existing support for downloading archives from CI builds.  The
PR id is translated into the latest successful CI run id for that PR.