SC2-forecasts: update README
Despite what the documentation says, `git subtree pull --squash`
does not squash commits. To do this we have to use `git subtree merge`
which is a little less user friendly but produces a nicer git history
for this repo, especially when we consider that the vendoring of
forecasts-ncov is temporary.