terraform: Unnest our cognito and iam modules
Nested modules run counter to the recommended patterns of module
composition¹, but it didn't pose issues when we had a single root
configuration.  Now that we're going to have more than one
configuration, it makes expressing optional module usage a lot harder.
(Just like the docs warned me of!)

The "moved" blocks let the old resource names in the current state
migrate to the new resource names when `terraform apply` is run.

¹ <https://developer.hashicorp.com/terraform/language/modules/develop/composition>