team: Align avatars to the top of the name/blurb block when present
This produces a more consistent visual repetition between blocks for
each person since the avatar's vertical positioning for each person
doesn't depend on the length/height of the blurb.  It also keeps the
avatar closer to the emphasized name it belongs to.  When scanning the
list of people, the avatars are what I key on first, and then I look
right to read the name/blurb.  With the previous styles, the name/blurb
was often back *up* the page a line height or two.