team: Use a proper grid layout instead of imitating one with flexbox
The goal of "team: Regularize avatar container widths" (7397d2d) was to
establish a regular grid, so do that with a grid layout not flexbox.

Using a flexbox to make a grid can get weird in edge cases of element
width/height and viewports because it packs elements differently than a
grid.