Allow read-only CORS requests for public resources
This makes it possible, for example, to do client-side analyses on the
dataset JSONs using standard nextstrain.org URLs.  Requested by
@corneliusroemer¹, as it makes it much easier to load arbitrary trees
into Nextclade web.

It's also nice in principle to allow cross-origin requests for public
resources since it enables unforeseen integrations without having to ask
permission.

I didn't use the "cors" library here since it does a lot more than what
we need, is more permissive by default, and setting our own headers is
less of a blackbox and thus clearer as to what we're sending and when.

¹ <https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1666027479770509>
  <https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1676473418174269>