Import forecast viz components (proof of principle)
This uses the approach first implemented in
https://github.com/nextstrain/forecasts-ncov/commit/7615f88f85948981cb7a68fa118105ea85932db0
(See https://github.com/nextstrain/forecasts-viz/commit/ba063b78668fee0e3f5bad89c3a2de47256e4d42
for a summary of the viz components we import here)

This approach would not work in our gatsby setup, resulting in a
separate PWA being used (with the server handling the routing) - a
generally poor outcome. It almost worked out-of-the box with Preact
(a small shim was needed).