Fix Gatsby proxy for dev
Prior to this commit, I was unable to load any individual Group page
after starting the dev server with `npm run dev`. I would get the error:
```
TypeError: (intermediate value).createProxyServer is not a function
```

Since `http-proxy` only has a default export¹, so we must access
the `default` property of the module in order to access
the `createProxyServer` method.

¹ https://github.com/http-party/node-http-proxy/blob/master/index.js#L13