Add `EditLogoForm` component
Adds group logo form with the options to delete the current group logo
or upload a new group logo. The form includes a display of the current
logo if it exists and a preview of the new logo before users upload
the new logo.

Future improvements can include a "confirmation" modal for deleting
or uploading logos.