Add OPTIONS endpoint for group settings
Returns the allowed HTTP methods for a group's settings routes based
on the request user's permissions for the group.

Intended to be used on the client side to check if the user is
authorized to see a "Edit Group Settings" button.

This commits includes an addition of a `.optionsAsync()` method for
consistency even though this is not an async endpoint.