Add basic framework for edit group settings page.
When on a groups page, users who have permission to edit group settings
will see a "Edit Group Settings" button in the top right corner that
will direct them to the new page.

Users can also directly navigate to the page with the URL
`/groups/:groupName/settings`. If the user is not logged in, this will
direct them to the login page. Users who do _not_ have permissions to
edit the group settings will see an error banner. This is mostly
implemented to catch cases where the user's role was updated but
their user info is "stale".