SC2-forecasts: Textual and stylistic pass
This commit adjusts text throughout the page as well as styling. I generally don't like readability of centered text with ragged edges. Moved to left justified, but have kept the paragraph from being over wide on large displays.

I've removed verbiage in the top description for the Rt renewal model, but have not yet removed the panels beneath titled smoothedIncidencePanel and rtPanel, but would plan to remove these.

I included the "enabled by GISAID" with the GISAID logo as is required for webapps.