Add `EditOverviewForm` component
Adds a new form to allow editing of the group overview in a textarea.
If the group does not already have an overview or their overview is
empty, then the textarea will display a default overview template.

Future improvements can include adding format validation prior to
upload and adding a preview of the overview.