edit-overview-form: Polish up the doc link
Switches to the /page/… alias for docs.nextstrain.org to future proof
against changes to default language (en; unlikely) and default version
(latest; more likely).

Removes the noopener and noreferrer relationships, as they're intended
for third-party-provided links and not necessary for our
implicitly-trusted docs site.  For a couple years now, browsers have
also treated target=_blank links as implicitly rel=noopener by default.
Finally, noreferrer implies noopener, so both were redundant.

Removes trailing space from the end of the link, which I believe was in
place due to a quirk of JSX, and replaces it with the more common
workaround of interpolating a space.  This makes the link underline less
unsightly by ending it with the last word rather than a space.