Install Next.js 13.3.0
We're going to use the beta "app router" feature in version 13 for
server-side rendering, integrated into our existing server.  Since it's
a beta feature that's explicitly not subject to semantic versioning
rules, we pin to an exact version (the current latest, 13.3.0).
Upgrades can and should be done with new releases, but they should be
handled manually after review of changes.