sources: Remove the abstract Subresource.validTypes static property getter
This caused an error in the Next.js build (when included by import)
because the validTypes property is redefined by subclasses.  The
redefinition is—due to transpilation, I think—treated as as a property
value assignment operation (i.e. attempt to call a setter that doesn't
exist) instead of a property definition (i.e. re-definition of the
existing property).

The abstract getter is only a guard against programming mistakes and not
strictly necessary, so remove it.  It's a little unfortunate since it
does help document the abstract Subresource interface, but I don't feel
like digging into why Webpack has an issue here at the moment.  Node.js
and browsers are seemingly fine with the construction we use in a toy
example.