Explore navigation: Link vs anchor elements
Sets up three pages, each one with Next.js Link elements and HTML anchor
elements in the page, allowing us to test the different styles of page
navigation on offer. Added URLs:
/parent
/parent/child1
/parent/child2

**Behaviour using Link elements**

Navigating from Parent to Child1 is
client-side, with a pre-fetch for some child1 data (text/x-component)
when we hover over the Link element. On clicking the Link we get some
more child1 data (text/x-component), which includes the elements to be
rendered on that page. Navigating back to the Parent page (via Link
elements) is done without _any_ network requests.

This behaviour is unchanged if one is logged in or not - note that every
(server-side) rendering of the NavBar, present on all of these pages,
calls `const currentUser = await getCurrentUser();`.

There are niceties about this style of client-side navigation,
especially if we have lots of JS in the page as (even if cached) it
needs to be re-parsed each time we change page.

**Behaviour using anchor (a) elements**

This is more traditional browser navigation behaviour. Each navigation
between pages makes a fresh batch of requests, including for the base
HTML page to be rendered. This results in the NavBar flashing, as it's
"new" (in the eyes of the browser). These flashes are improved when the
browser cache is used (which is our expected operating environment).

new file:   src/app/parent/child2/page.jsx #	new file:
src/app/parent/counter.jsx #	new file:   src/app/parent/page.jsx #	new
file:   src/app/parent/styles.css #