eslint: Address quote-props violations
The only violations are in the Jest config, which was previously ignored
by ESLint. If this is going to be a JS file, might as well lint it and
make it conform to JS file standards in this repo.

The official Jest docs do not use unnecessary quotes in their JS config
examples¹, which aligns with the fix here.

¹ https://jestjs.io/docs/configuration