eslint: Partially address no-use-before-define violations
Addressing the violations where a function expression is used before it
is defined. Changing to a function declaration solves this because those
are hoisted to the top-level (function expressions are not).

Disable this rule for classes - somehow those are working fine as-is.

Rule doc page: https://eslint.org/docs/latest/rules/no-use-before-define