Don't use Gatsby's default ESLint rules
To keep things simple, the existing top-level ESLint configuration is
intended to apply for the entire repository.

5c0f68f49607a4376b2e66496c3019fad9ec396c removed all per-line exceptions
in static-site/ under the assumption that ESLint was never run within
that directory due to ignore patterns. I now realize that is not
entirely true.

By default, Gatsby uses a predefined ESLint configuration¹. It has been
used here for a while, most likely unintentionally since static-site/ is
ignored in the top-level config.

Override it with an empty configuration file in the Gatsby project
root.

Notes:

1. Gatsby only runs ESLint during develop² and not build, which is why
   this wasn't caught by CI.
2. Gatsby will still run ESLint and add its own custom rules³.
   This empty file just prevents the default config from being used.

¹ https://www.gatsbyjs.com/docs/how-to/custom-configuration/eslint/#how-to-use-eslint
² https://github.com/gatsbyjs/gatsby/blob/gatsby@2.32.13/packages/gatsby/src/utils/webpack.config.js#L338-L355
³ https://github.com/gatsbyjs/gatsby/blob/gatsby@2.32.13/packages/gatsby/src/utils/webpack.config.js#L349-L355