WIP Client side code
This prototype is working as expected, but needs some improvements
before deployment:

* Bundling should use minimisation etc (and behave differently in dev mode)
* React & ReactDOM should be their own bundle as all of our current
  client-side code uses them. They can be loaded (via bootstrapModules)
  for pages as necessary.
* ./develop.sh should watch for changes to client-code

Bundling was used (as opposed to just using ESM) as modules are hard
to come by for many of our libraries. There's nothing stopping us moving
in that direction in the future. We can experiment as-is with this by
pointing the 'clientComponent' property to our own ESM file rather than
the output of a bundler.

Hydration, rather than the client code taking ownership of an ~empty
element can also be implemented as needed in the future.