Remove dataset listing /influenza, /staging
The JSON backing the influenza listing was empty, and the JSON backing
the staging dataset was extremely incomplete (29 datasets total).
I presume this is due to the truncated results of listing the bucket.
Rather than attempt to fix the script, I think it's advisable to revisit
these concepts and what we really want to do here. While this is
happening it's better to remove the incomplete list of datasets.

These pages were implemented with the hope