Implement filtering for getAvailable API
API design is WIP - see commentary added in this commit

There is subtlety in how we want prefix-based filtering to work, as the
_name_ of a set of datasets is not always the same as the filename,
as we have the ability to group multiple filenames (keys) together.
E.g. given the following collection of datasets:
  - name: X_Y
  - versions: [
    - name: X_Y (current version)
    - name: X_Y (old version)
    - name: X_Y_2022-01-01 (current version, datestamped filename)
  ]
Then prefix-based filtering on "X/Y/2022-01-01" could either
return nothing as it doesn't match the name assigned to the collection,
or return the dataset collection restricted to 1/3 of the versions.
Here I implement the former.