Recollect resources every ~24 hours
This uses a simple interval-based approach, which should be enough for
the core & staging sources.

Ideally we would re-fetch the inventory shortly after it's creation.
The time of day at which a new inventory appears seems random, so this
would require using events or polling the inventory directory more
frequently and only updating when a new manifest appears.
The directory name of inventory manifests seems to be consistently
YYYY-MM-DD + "T01-00Z", so perhaps it's enough to make HEAD requests
for the upcoming day's data?

2023-07-12T01-00Z/