Update /pathogens page
Reflects the new cards UI as the main interface to the datasets. The
previous datasets+narratives table is restricted to narratives-only,
as these are not yet (or may never be) part of the new cards UI.

The colourful tiles are restricted to 6 pathogens, which on most screen
resolutions render as 3 columns x 2 rows. The removed pathogens are
not being regularly updated, and the ultimate aim is for datasets to be
highlighted through the new UI.