authn: Generalize support to ~any OIDC/OAuth2 IdP, not just AWS Cognito
OIDC is OpenID Connect 1.0, which is an identity/authentication protocol
layered on top of OAuth 2.0's authorization protocol.  AWS Cognito
implements OIDC/OAuth2 but our authn code hardcoded some assumptions
about Cognito specifically.  Undo that and parameterize and generalize
the code to work (in theory) with other OIDC identity providers (IdPs).
In practice, some additional changes may be necessary for specific other
IdPs, but as-is I can get this generalized authn code to work against a
test Azure AD IdP.

Outside of authn, there are still some other bits of the codebase which
require Cognito.  Those will be addressed in subsequent work.

This work is motivated by CDC AMD's efforts to host a copy of
nextstrain.org internally in order to avail themselves of Groups
internally.

Related-to: <https://github.com/nextstrain/private/issues/94>