Remove the nonsense that is utils.getGitHash()
This function is used in one place, to provide a "version" that's not
necessary.  Since the implementation is nonsense to begin with, just
remove it.

While reporting the current SHA can be useful in many cases, I don't see
how it's actually useful here.  It's been broken since it was added in
November 2018 and in 5 years no one's missed it.  In local development,
you can see what revision you're using easily.  On Heroku dynos, the app
slugs don't contain .git/ so it'd always be "unknown" anyway.  So that's
why I removed it instead of fixing it.

Fixing the code is doable, of course, but the correct approach is not to
read .git/HEAD and resolve its contents ourselves.  One example where
that approach fails is in a "detached HEAD" state (e.g. you've checked
out a specific commit, perhaps for testing), but there are others.
Using `git rev-parse HEAD` or similar (e.g. `git describe --always
--dirty`) would be a better fix.  But again, it doesn't seem worth
bothering.