Parse resources index
Parses a pre-computed index JSON and stores the data in-memory (on the
nextstrain.org server). ResourceVersions is a class which
dataset/narrative requests can use to get the the (versioned) file URLs
for available subresources which were present for a given YYYY-MM-DD. A
subsequent commit will allow the usage of "@YYYY-DD-MM" descriptors in
URLs, and eventually these data will be used to display all available
resources.

A subsequent commit in this PR will add documentation regarding the
RESOURCE_INDEX variable, but it's design was influenced by
<https://github.com/nextstrain/nextstrain.org/pull/719#discussion_r1385803346>.
Briefly, to use a local index set `RESOURCE_INDEX="./path/to/json"` and
to disable the index use `RESOURCE_INDEX="false"`.