Allow versioned resource access (core/staging)
Nextstrain URLs are extended to allow <path>@<version> syntax for core
datasets. Currently the <version> must be in YYYY-MM-DD format. The
returned version is the one which was the latest on the requested day.
If the requested version predates any datasets we return 404.

Note that the current implementation (via previous commits) uses S3
versioning, not our datestamped datasets, although the two concepts may
appear similar in the URL.