Allow versioned resource access (core only)
Nextstrain URLs are extended to allow <path>@<version> syntax for core
datasets. Currently the <version> must be in YYYY-MM-DD format. The
returned version is the one which was the latest on the requested day.
If the requested version predates any datasets we return 404.

We attempt to extract a version descriptor for every request, however
for non-core sources (and core narratives)¹ the presence of such a
descriptor (i.e. if "@" is in the URL²) will result in a 400 (Bad
Request) error. For further examples please see
`test/date_descriptor.test.js`.

Note that the @YYYY-MM-DD URLs enabled by this commit look similar to
some existing URLs where the datestamp is in the dataset name (e.g.
"ncov/gisaid/global/6m/2024-01-02") however conceptually these are quite
different.

¹ Fetch URLs (e.g. /fetch/...) are excluded
² There are exceptions, such as community URLs allowing <repo>@<commit>
in the pathname.