static-site/README: add docs on how to add new team member
I could not find any docs on how to add new members to the team page.
I also had to follow through a linked GitHub commit¹ to
figure out how the footer in docs.nextstrain.org was updated the last
time we updated the team page² so adding those instructions here as well.

¹ https://github.com/nextstrain/sphinx-theme/commit/65487653771efbcfa19dbcf2adab19a90dd85318
² https://github.com/nextstrain/nextstrain.org/pull/640