redis: Require REDIS_URL in production mode (except for review apps)
Redis is our durable store for sessions and also user staleness markers.
These are both increasingly important for production use.  For example,
the API endpoints for modifying group members/roles relies on the user
staleness markers to force refreshing of cached info in the session of
an affected user.  This ensures membership/role changes are effectively
immediately instead of only after a delay (or after logout/login).

Rather than assuming Redis will always be configured in production,
explicitly assert that it is (similar to our assertion for the related
SESSION_ENCRYPTION_KEYS).

Review apps are exempt because due to how we configure them at the
moment we cannot automatically provision separate Redis instances for
them.