config: Configurability of the authn SESSION_* constants
Motivated by CDC's need to change the values of these constants for
production instances of the codebase which _aren't_ our own (i.e. which
are not nextstrain.org and next.nextstrain.org).  At the very least
SESSION_COOKIE_DOMAIN needed to be configurable, so it made sense to
also allow the related SESSION_SECRET to be set via config file (instead
of just env) and SESSION_MAX_AGE to be set too.

Our important production value for SESSION_COOKIE_DOMAIN is moved into
our production Terraform configuration as an output so that its included
in env/production/config.json with the rest of the non-sensitive values.