tests: Isolate process.env for each test in files where we modify it
Previously these tests were unintentionally relying on test order: the first
test—checking that SESSION_ENCRYPTION_KEYS is required in production—was
setting NODE_ENV=production and SESSION_ENCRYPTION_KEYS=… which leaked
into the subsequent tests.  This works… but it makes the tests more
fragile to have that coupling and the coupling was unintended.

I noticed this while changing the location of the SESSION_SECRET
requirement check from src/authn/index.js to src/config.js.  The latter
is loaded by src/authn/session.js, which is loaded by this test file,
while the former was not.  This made SESSION_SECRET a new requirement
for the first test here, and while I expected the first test to fail,
*all* the tests actually failed.  That led to the discovery that
NODE_ENV=production was leaking.